chagnes
thankss

+++ project/admin_products.php
@@ -6,74 +6,88 @@ session_start();
 
 $admin_id = $_SESSION['admin_id'];
 
-if(!isset($admin_id)){
+if (!isset($admin_id)) {
    header('location:login.php');
+   exit;
 };
 
-if(isset($_POST['add_product'])){
+if (isset($_POST['add_product'])) {
 
-   $name = mysqli_real_escape_string($conn, $_POST['name']);
+   $name = $_POST['name'];
    $price = $_POST['price'];
    $image = $_FILES['image']['name'];
    $image_size = $_FILES['image']['size'];
    $image_tmp_name = $_FILES['image']['tmp_name'];
-   $image_folder = 'uploaded_img/'.$image;
+   $image_folder = 'uploaded_img/' . $image;
 
-   $select_product_name = mysqli_query($conn, "SELECT name FROM `products` WHERE name = '$name'") or die('query failed');
+   $select_product_name = $conn->prepare("SELECT name FROM `products` WHERE name = :name");
+   $select_product_name->execute([':name' => $name]);
 
-   if(mysqli_num_rows($select_product_name) > 0){
-      $message[] = 'product name already added';
-   }else{
-      $add_product_query = mysqli_query($conn, "INSERT INTO `products`(name, price, image) VALUES('$name', '$price', '$image')") or die('query failed');
+   if ($select_product_name->rowCount() > 0) {
+      $message[] = 'Product name already added';
+   } else {
+      $add_product_query = $conn->prepare("INSERT INTO `products` (name, price, image) VALUES (:name, :price, :image)");
+      $add_product_query->execute([':name' => $name, ':price' => $price, ':image' => $image]);
 
-      if($add_product_query){
-         if($image_size > 2000000){
-            $message[] = 'image size is too large';
-         }else{
+      if ($add_product_query) {
+         if ($image_size > 2000000) {
+            $message[] = 'Image size is too large';
+         } else {
             move_uploaded_file($image_tmp_name, $image_folder);
-            $message[] = 'product added successfully!';
+            $message[] = 'Product added successfully!';
          }
-      }else{
-         $message[] = 'product could not be added!';
+      } else {
+         $message[] = 'Product could not be added!';
       }
    }
 }
 
-if(isset($_GET['delete'])){
+if (isset($_GET['delete'])) {
    $delete_id = $_GET['delete'];
-   $delete_image_query = mysqli_query($conn, "SELECT image FROM `products` WHERE id = '$delete_id'") or die('query failed');
-   $fetch_delete_image = mysqli_fetch_assoc($delete_image_query);
-   unlink('uploaded_img/'.$fetch_delete_image['image']);
-   mysqli_query($conn, "DELETE FROM `products` WHERE id = '$delete_id'") or die('query failed');
+
+   $delete_image_query = $conn->prepare("SELECT image FROM `products` WHERE id = :id");
+   $delete_image_query->execute([':id' => $delete_id]);
+   $fetch_delete_image = $delete_image_query->fetch(PDO::FETCH_ASSOC);
+
+   if ($fetch_delete_image) {
+      unlink('uploaded_img/' . $fetch_delete_image['image']);
+   }
+
+   $delete_product_query = $conn->prepare("DELETE FROM `products` WHERE id = :id");
+   $delete_product_query->execute([':id' => $delete_id]);
+
    header('location:admin_products.php');
+   exit;
 }
 
-if(isset($_POST['update_product'])){
+if (isset($_POST['update_product'])) {
 
    $update_p_id = $_POST['update_p_id'];
    $update_name = $_POST['update_name'];
    $update_price = $_POST['update_price'];
 
-   mysqli_query($conn, "UPDATE `products` SET name = '$update_name', price = '$update_price' WHERE id = '$update_p_id'") or die('query failed');
+   $update_product_query = $conn->prepare("UPDATE `products` SET name = :name, price = :price WHERE id = :id");
+   $update_product_query->execute([':name' => $update_name, ':price' => $update_price, ':id' => $update_p_id]);
 
    $update_image = $_FILES['update_image']['name'];
    $update_image_tmp_name = $_FILES['update_image']['tmp_name'];
    $update_image_size = $_FILES['update_image']['size'];
-   $update_folder = 'uploaded_img/'.$update_image;
+   $update_folder = 'uploaded_img/' . $update_image;
    $update_old_image = $_POST['update_old_image'];
 
-   if(!empty($update_image)){
-      if($update_image_size > 2000000){
-         $message[] = 'image file size is too large';
-      }else{
-         mysqli_query($conn, "UPDATE `products` SET image = '$update_image' WHERE id = '$update_p_id'") or die('query failed');
+   if (!empty($update_image)) {
+      if ($update_image_size > 2000000) {
+         $message[] = 'Image file size is too large';
+      } else {
+         $update_image_query = $conn->prepare("UPDATE `products` SET image = :image WHERE id = :id");
+         $update_image_query->execute([':image' => $update_image, ':id' => $update_p_id]);
          move_uploaded_file($update_image_tmp_name, $update_folder);
-         unlink('uploaded_img/'.$update_old_image);
+         unlink('uploaded_img/' . $update_old_image);
       }
    }
 
    header('location:admin_products.php');
-
+   exit;
 }
 
 ?>
@@ -84,7 +98,7 @@ if(isset($_POST['update_product'])){
    <meta charset="UTF-8">
    <meta http-equiv="X-UA-Compatible" content="IE=edge">
    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-   <title>products</title>
+   <title>Courses</title>
 
    <!-- font awesome cdn link  -->
    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css">
@@ -101,14 +115,14 @@ if(isset($_POST['update_product'])){
 
 <section class="add-products">
 
-   <h1 class="title">shop products</h1>
+   <h1 class="title">Shop Courses</h1>
 
    <form action="" method="post" enctype="multipart/form-data">
-      <h3>add product</h3>
-      <input type="text" name="name" class="box" placeholder="enter product name" required>
-      <input type="number" min="0" name="price" class="box" placeholder="enter product price" required>
+      <h3>Add Product</h3>
+      <input type="text" name="name" class="box" placeholder="Enter product name" required>
+      <input type="number" min="0" name="price" class="box" placeholder="Enter product price" required>
       <input type="file" name="image" accept="image/jpg, image/jpeg, image/png" class="box" required>
-      <input type="submit" value="add product" name="add_product" class="btn">
+      <input type="submit" value="Add Product" name="add_product" class="btn">
    </form>
 
 </section>
@@ -122,21 +136,21 @@ if(isset($_POST['update_product'])){
    <div class="box-container">
 
       <?php
-         $select_products = mysqli_query($conn, "SELECT * FROM `products`") or die('query failed');
-         if(mysqli_num_rows($select_products) > 0){
-            while($fetch_products = mysqli_fetch_assoc($select_products)){
+         $select_products = $conn->query("SELECT * FROM `products`");
+         if ($select_products->rowCount() > 0) {
+            while ($fetch_products = $select_products->fetch(PDO::FETCH_ASSOC)) {
       ?>
       <div class="box">
          <img src="uploaded_img/<?php echo $fetch_products['image']; ?>" alt="">
          <div class="name"><?php echo $fetch_products['name']; ?></div>
          <div class="price">$<?php echo $fetch_products['price']; ?>/-</div>
-         <a href="admin_products.php?update=<?php echo $fetch_products['id']; ?>" class="option-btn">update</a>
-         <a href="admin_products.php?delete=<?php echo $fetch_products['id']; ?>" class="delete-btn" onclick="return confirm('delete this product?');">delete</a>
+         <a href="admin_products.php?update=<?php echo $fetch_products['id']; ?>" class="option-btn">Update</a>
+         <a href="admin_products.php?delete=<?php echo $fetch_products['id']; ?>" class="delete-btn" onclick="return confirm('Delete this product?');">Delete</a>
       </div>
       <?php
          }
-      }else{
-         echo '<p class="empty">no Courses added yet!</p>';
+      } else {
+         echo '<p class="empty">No Courses added yet!</p>';
       }
       ?>
    </div>
@@ -146,37 +160,32 @@ if(isset($_POST['update_product'])){
 <section class="edit-product-form">
 
    <?php
-      if(isset($_GET['update'])){
+      if (isset($_GET['update'])) {
          $update_id = $_GET['update'];
-         $update_query = mysqli_query($conn, "SELECT * FROM `products` WHERE id = '$update_id'") or die('query failed');
-         if(mysqli_num_rows($update_query) > 0){
-            while($fetch_update = mysqli_fetch_assoc($update_query)){
+         $update_query = $conn->prepare("SELECT * FROM `products` WHERE id = :id");
+         $update_query->execute([':id' => $update_id]);
+         if ($update_query->rowCount() > 0) {
+            while ($fetch_update = $update_query->fetch(PDO::FETCH_ASSOC)) {
    ?>
    <form action="" method="post" enctype="multipart/form-data">
       <input type="hidden" name="update_p_id" value="<?php echo $fetch_update['id']; ?>">
       <input type="hidden" name="update_old_image" value="<?php echo $fetch_update['image']; ?>">
       <img src="uploaded_img/<?php echo $fetch_update['image']; ?>" alt="">
-      <input type="text" name="update_name" value="<?php echo $fetch_update['name']; ?>" class="box" required placeholder="enter product name">
-      <input type="number" name="update_price" value="<?php echo $fetch_update['price']; ?>" min="0" class="box" required placeholder="enter product price">
+      <input type="text" name="update_name" value="<?php echo $fetch_update['name']; ?>" class="box" required placeholder="Enter product name">
+      <input type="number" name="update_price" value="<?php echo $fetch_update['price']; ?>" min="0" class="box" required placeholder="Enter product price">
       <input type="file" class="box" name="update_image" accept="image/jpg, image/jpeg, image/png">
-      <input type="submit" value="update" name="update_product" class="btn">
-      <input type="reset" value="cancel" id="close-update" class="option-btn">
+      <input type="submit" value="Update" name="update_product" class="btn">
+      <input type="reset" value="Cancel" id="close-update" class="option-btn">
    </form>
    <?php
          }
       }
-      }else{
+      } else {
          echo '<script>document.querySelector(".edit-product-form").style.display = "none";</script>';
       }
    ?>
 
 </section>
-
-
-
-
-
-
 
 <!-- custom admin js file link  -->
 <script src="js/admin_script.js"></script>
